+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Features
+
+- Emit `canvas_theme` (`light`/`dark`) in the generated snapshot sidecar metadata, derived from the Compose `@Preview` `uiMode` ([#1364](https://github.com/getsentry/sentry-android-gradle-plugin/pull/1364))
+
 ### Dependencies
 
 - Bump CLI from v3.6.0 to v3.6.1 ([#1367](https://github.com/getsentry/sentry-android-gradle-plugin/pull/1367))